move validateDomains to management.domain pkg to reuse
