+++ client/internal/config.go
@@ -20,6 +20,7 @@ import (
 	"github.com/netbirdio/netbird/client/internal/routemanager/dynamic"
 	"github.com/netbirdio/netbird/client/ssh"
 	mgm "github.com/netbirdio/netbird/management/client"
+	"github.com/netbirdio/netbird/management/domain"
 	"github.com/netbirdio/netbird/util"
 )
 
@@ -69,7 +70,7 @@ type ConfigInput struct {
 
 	BlockLANAccess *bool
 
-	DNSLabels []string
+	DNSLabels domain.List
 }
 
 // Config Configuration type
@@ -95,7 +96,7 @@ type Config struct {
 
 	BlockLANAccess bool
 
-	DNSLabels []string
+	DNSLabels domain.List
 
 	// SSHKey is a private SSH key in a PEM format
 	SSHKey string
@@ -495,8 +496,8 @@ func (config *Config) apply(input ConfigInput) (updated bool, err error) {
 
 	if input.DNSLabels != nil && !reflect.DeepEqual(config.DNSLabels, input.DNSLabels) {
 		log.Infof("updating DNS labels [ %s ] (old value: [ %s ])",
-			strings.Join(input.DNSLabels, " "),
-			strings.Join(config.DNSLabels, " "))
+			input.DNSLabels.SafeString(),
+			config.DNSLabels.SafeString())
 		config.DNSLabels = input.DNSLabels
 		updated = true
 	}
